requirements.txt
pytoshop
psd-tools
Pillow>=8.0.0
torch>=1.7.0
numpy>=1.19.0